Do not cast float to float in UAGX_Constraint1DofComponent::GetAngle

+++ Source/AGXUnreal/Private/Constraints/AGX_Constraint1DofComponent.cpp
@@ -42,7 +42,7 @@ namespace
 
 float UAGX_Constraint1DofComponent::GetAngle() const
 {
-	return static_cast<float>(Get1DOFBarrier(*this)->GetAngle());
+	return Get1DOFBarrier(*this)->GetAngle();
 }
 
 void UAGX_Constraint1DofComponent::CreateNativeImpl()
